fire version1
The capacity is very large, so the upload is not possible, so only the script code file is uploaded for backup. I will revise the project file directly after November 17th when you implement the water. Based on these files, the smudging of fire, the sound of fire, and the extinguishing of fire are all implemented, and the code does not need to be supplemented, but the numbers still need to be adjusted.